fix decorator
